Add some platform detection to the Makefile
We need some platform-specific defines that are used by the Lua source
in e.g. luaconf.h. These would normally be set by the upstream Makefile,
but we're not using that. This commit checks the `uname_S` make variable
exported by the crawl Makefile set defines for Linux and MacOs.

MSVC compilation doesn't use a Makefile, and I'm not sure if the MSVC
project file needs updating to set relevant defines for 64-bit vs 32-bit
windows. For mingw targets, I think both native and cross-compilation
using the lua 5.4 contrib might be somewhat broken. However we currently
do cross-compilation with lua5.4 installed as a system package, and it's
likely a native msys2 environment also has a lua 5.4 package available.
So for now, I'm not going to worry about adapting the lua contrib
Makefile for mingw.